[WFCORE-452] Wire extension resources into host model
This is needed since in order to configure subsystems on a host the
extensions from the domain model have not yet been initialised.
Some logic to make sure that defining extensions in both the domain and
host model is not an error, and to make sure that only the last remove
from the model does the actual removal from the registry.